Remove cyclic dependency
